API review: move version module to avocado.core
While still making the VERSION symbol available in the top level
Avocado module. Even though this is very simple information that
could live under the avocado module itself, it would break the
consistency of having avocado.core importing from avocado (toplevel).
